+++ cashu-lib-crypto/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+All notable changes to cashu-lib-crypto will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+---
+
+## [0.11.0] - 2026-01-10
+
+### Changed
+
+- Version sync with parent cashu-lib 0.11.0
+
+---
+
 ## 1.0.0 (2025-08-30)